+++ services/rds/delete_worker.go
@@ -2,13 +2,10 @@ package rds
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"log/slog"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/rds"
-	rdsTypes "github.com/aws/aws-sdk-go-v2/service/rds/types"
 	"github.com/cloud-gov/aws-broker/base"
 	"github.com/cloud-gov/aws-broker/config"
 	jobs "github.com/cloud-gov/aws-broker/jobs"
@@ -151,17 +148,4 @@ func (w *DeleteWorker) asyncDeleteDB(ctx context.Context, i *RDSInstance) error 
 
 	jobs.ShouldWriteAsyncJobMessage(w.db, i.ServiceID, i.Uuid, operation, base.InstanceGone, "Successfully deleted database resources")
 	return nil
-}
-
-func prepareDeleteDbInput(database string) *rds.DeleteDBInstanceInput {
-	return &rds.DeleteDBInstanceInput{
-		DBInstanceIdentifier:   aws.String(database), // Required
-		DeleteAutomatedBackups: aws.Bool(false),
-		SkipFinalSnapshot:      aws.Bool(true),
-	}
-}
-
-func isDatabaseInstanceNotFoundError(err error) bool {
-	var notFoundException *rdsTypes.DBInstanceNotFoundFault
-	return errors.As(err, &notFoundException)
 }
